refactor(ui): extract shared session frame renderer into session-frame.jsx
Moves SESSION_TERMINAL and the live-stream frame renderer + coalescing
helpers (_SLS_coalesceMessages/_SLS_NodeErrorBadge/_SLS_Frame/_SLS_ExpandableRow)
verbatim out of session-detail.jsx so the graph node inspector can render
node records identically. No behavior change; repoints two existing
session-detail grep-tests at the new module.
