+++ ui/index.html
@@ -79,6 +79,7 @@
 <script type="text/babel" src="components/workers.jsx"></script>
 <script type="text/babel" src="components/health.jsx"></script>
 <script type="text/babel" src="components/sessions-list.jsx"></script>
+<script type="text/babel" src="components/session-frame.jsx"></script>
 <script type="text/babel" src="components/session-detail.jsx"></script>
 <script type="text/babel" src="components/workspaces/shared.jsx"></script>
 <script type="text/babel" src="components/workspaces/providers.jsx"></script>
